requirements.txt
Flask
requests
gunicorn
flask-cors
PyJWT
